+++ src/Farmer/Arm/ContainerService.fs
@@ -187,7 +187,7 @@ type ManagedCluster = {
             OsDiskSize: int<Gb>
             OsType: OS
             VmSize: VMSize
-            AvailabilityZones: string list
+            AvailabilityZones: ZoneSelection
             VirtualNetworkName: ResourceName option
             SubnetName: ResourceName option
             PodSubnetName: ResourceName option
@@ -195,6 +195,7 @@ type ManagedCluster = {
             ScaleDownMode: ScaleDownMode option
             MinCount: int option
             MaxCount: int option
+            NodeTaints: string list option
         |} list
     DnsPrefix: string
     EnableRBAC: bool
@@ -214,7 +215,6 @@ type ManagedCluster = {
         {|
             NetworkPlugin: ContainerService.NetworkPlugin option
             DnsServiceIP: System.Net.IPAddress option
-            DockerBridgeCidr: IPAddressCidr option
             LoadBalancerSku: LoadBalancer.Sku option
             ServiceCidr: IPAddressCidr option
         |} option
@@ -300,7 +300,7 @@ type ManagedCluster = {
                                 osDiskSizeGB = agent.OsDiskSize
                                 osType = string agent.OsType
                                 vmSize = agent.VmSize.ArmValue
-                                availabilityZones = agent.AvailabilityZones
+                                availabilityZones = agent.AvailabilityZones.ArmValue
                                 vnetSubnetID =
                                     match agent.VirtualNetworkName, agent.SubnetName with
                                     | Some vnet, Some subnet -> subnets.resourceId(vnet, subnet).Eval()
@@ -316,6 +316,7 @@ type ManagedCluster = {
                                     | _ -> null
                                 minCount = agent.MinCount |> Option.toNullable
                                 maxCount = agent.MaxCount |> Option.toNullable
+                                nodeTaints = agent.NodeTaints
                             |})
                         dnsPrefix = this.DnsPrefix
                         enableRBAC = this.EnableRBAC
@@ -344,10 +345,6 @@ type ManagedCluster = {
                             match this.NetworkProfile with
                             | Some networkProfile -> {|
                                 dnsServiceIP = networkProfile.DnsServiceIP |> Option.map string |> Option.toObj
-                                dockerBridgeCidr =
-                                    networkProfile.DockerBridgeCidr
-                                    |> Option.map IPAddressCidr.format
-                                    |> Option.toObj
                                 loadBalancerSku =
                                     networkProfile.LoadBalancerSku
                                     |> Option.map (fun sku -> sku.ArmValue)