+++ src/Farmer/Builders/Builders.ContainerApps.fs
@@ -1,7 +1,9 @@
 [<AutoOpen>]
 module Farmer.Builders.ContainerApps
 
+open System
 open Farmer
+open Farmer.ApplicationGateway
 open Farmer.Builders
 open Farmer.ContainerApp
 open Farmer.ContainerAppValidation
@@ -18,6 +20,7 @@ type ContainerConfig = {
         Memory: float<Gb>
         EphemeralStorage: float<Gb> option
     |}
+    Probes: ProbeMap
 } with
 
     member internal this.BuildContainer: Container =
@@ -27,6 +30,7 @@ type ContainerConfig = {
             DockerImage = dockerImage
             Resources = this.Resources
             VolumeMounts = this.VolumeMounts
+            Probes = this.Probes
           }
         | None -> raiseFarmer $"Container '{this.ContainerName}' requires a docker image."
 
@@ -589,6 +593,7 @@ type ContainerAppBuilder() =
             DockerImage = Some(Containers.PublicImage(dockerImage, Some dockerVersion))
             Resources = defaultResources
             VolumeMounts = Map.empty
+            Probes = Map.empty
         }
 
         this.AddContainers(state, [ container ])
@@ -617,6 +622,7 @@ type ContainerBuilder() =
         DockerImage = None
         Resources = defaultResources
         VolumeMounts = Map.empty
+        Probes = Map.empty
     }
 
     /// Set docker credentials
@@ -690,6 +696,20 @@ type ContainerBuilder() =
             VolumeMounts =
                 mounts
                 |> Seq.fold (fun s (volumeName, mountPath) -> s |> Map.add volumeName mountPath) state.VolumeMounts
+    }
+
+    [<CustomOperation "set_probe">]
+    member _.SetHealthProbe(state: ContainerConfig, probe, protocol, route, port) = {
+        state with
+            Probes =
+                state.Probes.Add(
+                    probe,
+                    {|
+                        Protocol = protocol
+                        Route = Uri(route, UriKind.Relative)
+                        Port = port
+                    |}
+                )
     }
 
 type DaprComponentBuilder() =
@@ -865,7 +885,6 @@ type DaprComponentBuilder() =
 
 
 let containerEnvironment = ContainerEnvironmentBuilder()
-
 let containerApp = ContainerAppBuilder()
 let container = ContainerBuilder()
 let daprComponent = DaprComponentBuilder()