+++ src/Farmer/Builders/Builders.LoadBalancer.fs
@@ -16,6 +16,7 @@ type FrontendIpConfig = {
     AddressVersion: AddressVersion
     PublicIp: LinkedResource option
     Subnet: LinkedResource option
+    Zones: ZoneSelection
 } with
 
     static member BuildResource frontend = {|
@@ -24,6 +25,7 @@ type FrontendIpConfig = {
         AddressVersion = frontend.AddressVersion
         PublicIp = frontend.PublicIp |> Option.map (fun linkedRes -> linkedRes.ResourceId)
         Subnet = frontend.Subnet |> Option.map (fun linkedRes -> linkedRes.ResourceId)
+        Zones = frontend.Zones
     |}
 
     static member BuildIp
@@ -37,12 +39,9 @@ type FrontendIpConfig = {
                 Name = resId.Name
                 AllocationMethod = AllocationMethod.Static
                 AddressVersion = frontend.AddressVersion
-                AvailabilityZone = None
+                AvailabilityZones = frontend.Zones
                 Location = location
-                Sku =
-                    match lbSku with
-                    | Farmer.LoadBalancer.Sku.Basic -> PublicIpAddress.Sku.Basic
-                    | Farmer.LoadBalancer.Sku.Standard -> PublicIpAddress.Sku.Standard
+                Sku = Sku.Standard
                 DomainNameLabel = None
                 Tags = Map.empty
             }
@@ -57,6 +56,7 @@ type FrontendIpBuilder() =
         AddressVersion = AddressVersion.IPv4
         PublicIp = None
         Subnet = None
+        Zones = NoZone
     }
 
     /// Sets the name of the frontend IP configuration.
@@ -102,6 +102,28 @@ type FrontendIpBuilder() =
     member _.LinkToUnmanagedSubnet(state: FrontendIpConfig, subnetId) = {
         state with
             Subnet = Some(Unmanaged subnetId)
+    }
+
+    /// Adds availability zones for the frontend, and therefore, the load balancer itself.
+    [<CustomOperation "add_availability_zones">]
+    member _.AddAvailabilityZones(state: FrontendIpConfig, zones: string seq) =
+        let existingZones =
+            match state.Zones with
+            | ExplicitZones existing -> existing
+            | ZoneExpression _
+            | NoZone -> []
+
+        {
+            state with
+                Zones = Seq.append existingZones zones |> Set.ofSeq |> Set.toSeq |> ExplicitZones
+        }
+
+    /// Adds availability zones in regions where they are available.
+    [<CustomOperation "pick_zones">]
+    member _.PickZones(state: FrontendIpConfig) = {
+        state with
+            // Doesn't work with 'loadBalancers' so use the response for 'publicIpAddresses'
+            Zones = ArmExpression.pickZones (publicIPAddresses, numZones = 3) |> ZoneExpression
     }
 
 let frontend = FrontendIpBuilder()
@@ -527,7 +549,7 @@ type LoadBalancerBuilder() =
     member _.Yield _ = {
         Name = ResourceName.Empty
         Sku = {
-            Name = LoadBalancer.Sku.Basic
+            Name = LoadBalancer.Sku.Standard
             Tier = LoadBalancer.Tier.Regional
         }
         FrontendIpConfigs = []