+++ RELEASE_NOTES.md
@@ -1,8 +1,20 @@
 Release Notes
 =============
 
+## 1.9.16
+* Container Apps: Adds `add_key_vault_secret` allowing Container App secrets to reference secrets stored in Azure Key Vault.
+
+## 1.9.15
+* Container Apps: Adds health probes.
+* AKS: Add node_taints parameter to agent pool config.
+
+## 1.9.14
+* VMSS, AKS, Azure Firewall, Managed Clusters (AKS): Adds `pick_zones` to let ARM select the availability zones.
+* AKS: Remove deprecated `docker_bridge` from Container Services (AKS).
+
 ## 1.9.13
-* Container Apps: Able to create Container Apps secrets referencing KeyVault secrets.
+* Subnets: Support for `disable_outbound_access`.
+* Public IP Addresses and Load Balancers: drop `Basic` SKU, support zones.
 
 ## 1.9.12
 * AKS Cluster: Add link_to_kubelet_identity